moved encrypt and decrypt functions to utils.
